% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{stacktrace}
\alias{stacktrace}
\title{Extract the current context stack to a data frame.}
\usage{
stacktrace()
}
\value{
A data frame with one row per stack frame, with deeper frames
  first. The stack frame that executes "context" is included, as well
  as context frames corresponding to builtins, looping and such.
}
\description{
Extract the current context stack to a data frame.
}

